requirements.txt
streamlit
openai
sentence-transformers
PyMuPDF
matplotlib
python-dotenv
